requirements.txt
python-dotenv
langchain
# langchain-openai
mcp-use
# ollama-python
langchain-ollama